blood-agent-api/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
minio
openai

python-dotenv
pydantic-ai